feat: change subdomains handler.
