+++ api/src/services/helper/addNewRoute.js
@@ -1,19 +1,9 @@
 import Router from 'koa-router';
+import handler from './subdomainHandler';
 
 const router = new Router();
 
 export const addNewRoute = ({ method, statusCode, body, path, headers }) => {
-  const handler = async ctx => {
-    headers.forEach(header => {
-      Object.keys(header).forEach(key => {
-        ctx.set(key, header[key]);
-      });
-    });
-
-    ctx.status = statusCode;
-    ctx.body = body;
-  };
-
   let newRoute;
   switch (method) {
     case 'GET':
